Fix duplicate logs
